op-e2e: Migrate withdrawals and fees into their own package
Continues the process of migrating e2e tests into discrete packages so they can be parallelized.
